requirements.txt
httpx
psycopg2-binary
python-dotenv
anthropic
rich
easyocr
mss
numpy
Pillow
opencv-python
